+++ src/BiotSavart/pointdata.jl
@@ -100,7 +100,7 @@ function add_point_charges!(data::PointData, fs::AbstractVector{<:AbstractFilame
     Ncharges = _count_charges(quad, fs)
     set_num_points!(data, Ncharges)
     inds = ChunkSplitters.chunks(eachindex(fs); n = Threads.nthreads())
-    Threads.@threads :static for subinds ∈ inds
+    Threads.@threads for subinds ∈ inds
         subinds === nothing && continue  # subinds can be nothing; see iterate(c::Chunk) in ChunkSplitters
         prev_indices = firstindex(fs):(first(subinds) - 1)  # filament indices given to all previous chunks
         n = _count_charges(quad, view(fs, prev_indices))    # we will start writing at index n + 1