+++ src/Forcing/Forcing.jl
@@ -21,11 +21,13 @@ abstract type AbstractForcing end
 """
     Forcing.apply!(forcing::AbstractForcing, vs::AbstractVector{<:Vec3}, f::AbstractFilament)
 
-Apply forcing to a single filament `f`, modifying vortex velocities `vs`.
+Apply forcing to a single filament `f` with self-induced velocities `vs`.
+
+At output, the `vs` vector is overwritten with the actual vortex line velocities.
 
 ---
 
-    Forcing.apply!(forcing::NormalFluidForcing, vs::AbstractVector{<:Vec3}, vn::AbstractVector{<:Vec3}, tangents::AbstractVector{<:Vec3})
+    Forcing.apply!(forcing::NormalFluidForcing, vs, vn, tangents)
 
 This variant can be used in the case of a [`NormalFluidForcing`](@ref) if one already has
 precomputed values of the normal fluid velocity and local unit tangents at filament points.
@@ -48,10 +50,10 @@ In particular, the function could be a synthetic velocity field from the
 [`SyntheticFields`](@ref) module (see below for examples).
 
 This type of forcing defines an external velocity ``\bm{v}_{\text{f}}`` affecting vortex
-motion, so that the effective vortex velocity is
+motion, so that the actual vortex velocity ``\bm{v}_{\text{L}}`` is
 
 ```math
-\frac{\mathrm{d}\bm{s}}{\mathrm{d}t} = \bm{v}_{\text{s}} + \bm{v}_{\text{f}}.
+\frac{\mathrm{d}\bm{s}}{\mathrm{d}t} = \bm{v}_{\text{L}} = \bm{v}_{\text{s}} + \bm{v}_{\text{f}}.
 ```
 
 Here ``\bm{v}_{\text{s}}`` is the self-induced velocity obtained by applying Biot–Savart's law.
@@ -137,7 +139,7 @@ function get_velocities!(forcing::NormalFluidForcing, vn::AbstractVector, f::Abs
 end
 
 function apply!(forcing::NormalFluidForcing, vs::AbstractVector, f::AbstractFilament)
-    length(vs) == length(f) || throw(DimensionMismatch("lengths of filament and velocity vector don't match"))
+    eachindex(vs) == eachindex(f) || throw(DimensionMismatch("lengths of filament and velocity vectors don't match"))
     @inline get_at_node(i) = (v⃗ₙ = forcing.vn(f[i]), s⃗′ = f[i, UnitTangent()])
     _apply!(get_at_node, forcing, vs)
 end
@@ -149,21 +151,22 @@ function apply!(forcing::NormalFluidForcing, vs::AbstractVector, vn::AbstractVec
 end
 
 # The first argument is a `get_at_node(i)` function which returns a NamedTuple with fields
-# (v⃗ₙ, s⃗′) with the normal fluid velocity and the local unit tangent at the node `i` of a
-# filament. This is useful if those quantities have been precomputed.
+# (v⃗ₛ, v⃗ₙ, s⃗′) with the superfluid velocity, normal fluid velocity and the local unit
+# tangent at the node `i` of a filament. This is useful if those quantities have been
+# precomputed.
 function _apply!(get_at_node::F, forcing::NormalFluidForcing, vs::AbstractVector) where {F <: Function}
     (; α, α′,) = forcing
     V = eltype(vs)  # usually Vec3{T} = SVector{3, T}
     for i ∈ eachindex(vs)
-        v⃗ₛ = vs[i]  # vortex velocity before mutual friction forcing
-        (; v⃗ₙ, s⃗′,) = @inline get_at_node(i)  # local normal fluid velocity and unit tangent
-        vₙₛ = V(v⃗ₙ) - v⃗ₛ     # slip velocity
+        (; v⃗ₙ, s⃗′,) = @inline get_at_node(i)  # superfluid velocity, normal fluid velocity and unit tangent
+        v⃗ₛ = vs[i]
+        vₙₛ = V(v⃗ₙ) - V(v⃗ₛ)   # slip velocity
         v_perp = s⃗′ × vₙₛ
-        vf = α * v_perp      # velocity due to Magnus force
+        vf = α * v_perp    # velocity due to Magnus force
         if !iszero(α′)
             vf = vf - α′ * s⃗′ × v_perp  # velocity due to drag force (it's quite common to set α′ = 0)
         end
-        vs[i] = vs[i] + vf
+        vs[i] = v⃗ₛ + vf
     end
     vs
 end