+++ src/Timestepping/forcing.jl
@@ -50,29 +50,40 @@ end
 
 function apply_forcing!(fields::NamedTuple, iter::VortexFilamentSolver, fs, time, to)
     if haskey(fields, :velocity)
+        # At this point, fields.velocity is expected to have the self-induced velocity vs.
+        # After applying a normal fluid forcing, fields.velocity will contain the actual
+        # vortex velocity vL. The self-induced velocity will be copied to iter.quantities.vs.
         _apply_forcing!(fields.velocity, iter.forcing, iter, fs, time, to)
     end
     fields
 end
 
-_apply_forcing!(vs_all, ::Nothing, args...) = nothing  # do nothing
+_apply_forcing!(vL, ::Nothing, args...) = nothing  # do nothing
 
-function _apply_forcing!(vs_all, forcing::NormalFluidForcing, iter, fs, t, to)
-    vn_all = iter.vn
-    tangents_all = iter.tangents
-    @assert eachindex(vn_all) == eachindex(tangents_all) == eachindex(vs_all) == eachindex(fs)
+function _apply_forcing!(vL_all, forcing::NormalFluidForcing, iter, fs, t, to)
+    # Note: inside a RK substep, quantities.{vs,vn,tangents} are used as temporary buffers
+    # (and in fact we don't really need vs).
+    (; quantities,) = iter
+    vs_all = quantities.vs  # self-induced velocities will be copied here
+    vn_all = quantities.vn  # normal fluid velocities will be computed here
+    tangents_all = quantities.tangents  # local tangents
+    @assert vs_all !== vL_all
+    @assert eachindex(vL_all) == eachindex(vn_all) == eachindex(tangents_all) == eachindex(vs_all) == eachindex(fs)
     @timeit to "Add forcing" begin
-        @inbounds Threads.@threads for n ∈ eachindex(fs, vs_all, vn_all, tangents_all)
+        @inbounds Threads.@threads for n ∈ eachindex(fs)
             f = fs[n]
             vs = vs_all[n]
+            vL = vL_all[n]
             vn = vn_all[n]
             tangents = tangents_all[n]
-            # Compute tangents (TODO: can we reuse them?)
+            # At input, vL contains the self-induced velocity vs.
+            # We copy vL -> vs before modifying vL with the actual vortex velocities.
             for i ∈ eachindex(f, tangents)
-                tangents[i] = f[i, UnitTangent()]
+                tangents[i] = f[i, UnitTangent()]  # TODO: can we reuse the tangents / computed elsewhere?
+                vs[i] = vL[i]
             end
             Forcing.get_velocities!(forcing, vn, f)    # evaluate normal fluid velocities
-            Forcing.apply!(forcing, vs, vn, tangents)  # modify vs according to the given forcing
+            Forcing.apply!(forcing, vL, vn, tangents)  # compute vL according to the given forcing (vL = vs at input)
         end
     end
     nothing