+++ CHANGELOG.md
@@ -5,7 +5,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- **Breaking**: When a normal fluid is applied, the `vs` field of
+  a `VortexFilamentSolver` (usually `iter.vs`) now represents the self-induced
+  vortex velocity, while the `vL` field is the actual filament velocity
+  including the contribution from mutual friction.
+
+- Short-range parallelisation using threads: use finer-grained parallelisation
+  at the level of the nodes of each filament. This is fine since the cost of
+  computing all short-range interactions on a filament node is relatively large,
+  so that the overhead of threading is relatively very small. Moreover, it's
+  much faster when we have a small number of filaments (which, counterintuitively,
+  can be relevant for turbulent cases). Also, we now avoid using `@threads :static`.
+
 ### Added
+
+- Store self-induced velocity ($v_s$) in addition to the actual filament
+  velocity ($v_L$) in `Timestepping.VortexFilamentSolver`.
 
 - Allow saving state of `FourierBandVectorField` to HDF5 file and loading it back.
   This is useful in particular for restarts.